+++ apps/server/src/docs/swagger.yaml
@@ -966,7 +966,7 @@ paths:
                   example: 1234
       responses:
         '200':
-          description: 주문 목록 조회 성공
+          description: (비회원) 주문 목록 조회 성공
           content:
             application/json:
               schema:
@@ -975,19 +975,11 @@ paths:
                   type: object
                   required:
                     - no
-                    - address
-                    - addressDetail
-                    - deliveryMethodName
                     - orderNumber
                     - orderStatus
-                    - orderStatusName
-                    - ordererMobile
-                    - ordererName
-                    - ordererEmail
-                    - recipientName
-                    - totalPrice
+                    - orderRoundNo
                     - createdAt
-                    - updatedAt
+                    - orderItems
                   properties:
                     no:
                       type: integer
@@ -995,70 +987,50 @@ paths:
                       example: 1
                     orderNumber:
                       type: string
-                      description: 주문 고유번호
-                      example: ORD-20240622-12345
+                      description: 주문 번호
+                      example: 'ORD20240825001'
                     orderStatus:
                       type: string
-                      enum: ['10', '20', '30', '40', '50', '51', '52']
-                      description: '주문 상태 (10-접수요청, 20-제조중, 30-배송중, 40-완료, 50-취소요청, 51-취소완료, 52-취소완료(환불))'
+                      description: 주문 상태 코드
                       example: '10'
                     orderStatusName:
                       type: string
                       description: 주문 상태명
-                      example: '접수요청'
-                    ordererMobile:
-                      type: string
-                      description: 주문자 전화번호
-                      example: '010-1234-5678'
-                    ordererName:
-                      type: string
-                      description: 주문자 이름
-                      example: '홍길동'
-                    ordererEmail:
-                      type: string
-                      description: 주문자 이메일
-                      example: 'test1234@naver.com'
-                    recipientName:
-                      type: string
-                      description: 수령인 이름
-                      example: '홍길동'
-                    totalPrice:
+                      example: '결제완료'
+                    orderRoundNo:
                       type: integer
-                      description: 총 주문 금액
-                      example: 15000
+                      description: 주문차수 번호
+                      example: 1
                     createdAt:
                       type: string
                       format: date-time
                       description: 주문 생성일시
-                      example: '2024-06-22T12:34:56.000Z'
-                    updatedAt:
-                      type: string
-                      format: date-time
-                      description: 주문 수정일시
-                      example: '2024-06-22T12:34:56.000Z'
-                    address:
-                      type: string
-                      description: "주소"
-                      example : "서울시 강남구"
-                    addressDetail:
-                      type: string
-                      description: "상세 주소"
-                      example : "123-45"
-                    deliveryMethodName:
-                      type: string
-                      description: "배송방법"
-                      example : "우체국"
-                    zipcode:
-                      type: string
-                      description: "우편번호"
-                      example : "12345"
-                    payment:
-                      type: object
-                      properties:
-                        isPaid:
-                          type: boolean
-                          description: 입금 확인 여부
-                          example: false
+                      example: '2024-08-25T15:55:20.000Z'
+                    orderItems:
+                      type: array
+                      items:
+                        type: object
+                        properties:
+                          no:
+                            type: integer
+                            description: 주문 아이템 번호
+                            example: 1
+                          breadName:
+                            type: string
+                            description: 상품명
+                            example: '식빵'
+                          breadImageUrl:
+                            type: string
+                            description: 상품 이미지 URL
+                            example: 'https://example.com/bread.jpg'
+                          unitPrice:
+                            type: number
+                            description: 단가
+                            example: 3000
+                          quantity:
+                            type: integer
+                            description: 수량
+                            example: 2
 
 
   /orders/{no}:
